requirements.txt
pandas
numpy
requests
black
Ipython
beartype
mypy
pylint
python-dotenv
openpyxl
beautifulsoup4
selenium
